+++ src/extension.ts
@@ -1,6 +1,7 @@
 import * as vscode from "vscode";
 import { registerCommands } from "./commands";
 import { registerCommentController } from "./comments";
+import * as config from "./config";
 import { registerFileSystemProvider } from "./fileSystem";
 import { registerRepoModule } from "./repos";
 import { extendMarkdownIt } from "./repos/wiki/markdownPreview";
@@ -9,12 +10,15 @@ import { store } from "./store";
 import { initializeAuth } from "./store/auth";
 import { initializeStorage } from "./store/storage";
 import { registerCodeSwingModule } from "./swings";
+import { Trace } from "./tracing";
 import { registerTreeProvider } from "./tree";
-import { registerProtocolHander } from "./uriHandler";
+import { registerProtocolHandler } from "./uriHandler";
+
+export let tracing: Trace;
 
 export async function activate(context: vscode.ExtensionContext) {
   registerCommands(context);
-  registerProtocolHander();
+  registerProtocolHandler();
   registerFileSystemProvider(store);
   registerTreeProvider(store, context);
   registerCommentController();
@@ -27,6 +31,15 @@ export async function activate(context: vscode.ExtensionContext) {
   registerShowcaseModule(context);
 
   const keysForSync = ["followedUsers", "repos"].map((key) => `gistpad.${key}`);
+  if (config.get("tracing.enableOutputChannel")) {
+    tracing = new Trace();
+  }
+
+  tracing?.appendLine(
+    `Setting keysForSync = ${keysForSync}`,
+    tracing.messageType.Info
+  );
+
   context.globalState.setKeysForSync(keysForSync);
 
   return {