+++ src/repos/store/storage.ts
@@ -1,4 +1,5 @@
 import { commands, ExtensionContext } from "vscode";
+import { tracing } from "../../extension";
 
 const HASREPOS_CONTEXT = "gistpad:hasRepos";
 const REPO_KEY = "gistpad.repos";
@@ -7,13 +8,25 @@ export interface IStorage {
   repos: string[];
 }
 
-export let storage: IStorage;
+export let reposStorage: IStorage;
 export async function initializeStorage(context: ExtensionContext) {
-  storage = {
+  reposStorage = {
     get repos(): string[] {
+      tracing?.appendLine(
+        `Getting repos from global state = ${context.globalState.get(
+          REPO_KEY,
+          []
+        )}`,
+        tracing?.messageType.Info
+      );
+
       return context.globalState.get(REPO_KEY, []);
     },
     set repos(repos: string[]) {
+      tracing?.appendLine(
+        `Setting repos to ${repos}`,
+        tracing?.messageType.Info
+      );
       context.globalState.update(REPO_KEY, repos);
       commands.executeCommand("setContext", HASREPOS_CONTEXT, repos.length > 0);
     }
@@ -22,6 +35,6 @@ export async function initializeStorage(context: ExtensionContext) {
   commands.executeCommand(
     "setContext",
     HASREPOS_CONTEXT,
-    storage.repos.length > 0
+    reposStorage.repos.length > 0
   );
 }