+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v0.5.0 (06/16/2022)
+
+- Fix sync across machines (for real this time)
+- Add optional Output trace channel
+  - `gistpad.tracing.enableOutputChannel`: default `false`. When enabled, creates an Output trace channel at VSCode startup
+  - `gistpad.tracing.automaticallyShowOutputChannel`: default `false`. When enabled, if `gistpad.tracing.enableOutputChannel` is also enabled, VSCode will automatically open and focus the GistPad Output channel
+
 ## v0.4.1 (04/25/2022)
 
 - Gists for followed users with more than 100 are now displayed properly (previously, the list was capped at the first 100 gists)